refactor : CoverImage 의 많은 필드의 원시값을 포장하여 관련된 것 끼리 분리
