+++ src/test/java/nextstep/courses/domain/CoverImageTest.java
@@ -15,29 +15,4 @@ class CoverImageTest {
         }).isInstanceOf(CanNotCreateException.class)
             .hasMessage("커버 이미지의 크기는 1MB 이상이어야 한다");
     }
-    
-    @Test
-    void 커버이미지의_정해진_크기_이하이면_에러전파() {
-        assertThatThrownBy(() -> {
-            new CoverImage(1_500_000, CoverImageType.JPEG, 299, 200);
-        }).isInstanceOf(CanNotCreateException.class)
-            .hasMessage("이미지의 너비(width)는 300px 이상이어야 한다");
-    }
-    
-    @Test
-    void 커버이미지의_정해진_비율이_아니면_에러전파() {
-        assertThatThrownBy(() -> {
-            new CoverImage(1_500_000, CoverImageType.JPEG, 300, 199);
-        }).isInstanceOf(CanNotCreateException.class)
-            .hasMessage("이미지의 높이(height)는 200px 이상이어야 한다");
-    }
-    
-    @Test
-    void 커버이미지의_너비와_높이_비율은_3대2가_아니면_에러전파() {
-        assertThatThrownBy(() -> {
-            new CoverImage(1_500_000, CoverImageType.JPEG, 300, 300);
-        }).isInstanceOf(CanNotCreateException.class)
-            .hasMessage("이미지의 너비(width)와 높이(height)의 비율은 3:2 이어야 한다");
-    }
-    
 }
